+++ src/org/lukasj/idea/torquescript/completion/TSCaseCorrectingInsertHandler.kt
@@ -1,0 +1,22 @@
+package org.lukasj.idea.torquescript.completion
+
+import com.intellij.codeInsight.completion.InsertHandler
+import com.intellij.codeInsight.completion.InsertionContext
+import com.intellij.codeInsight.lookup.LookupElement
+import com.intellij.openapi.editor.EditorModificationUtil
+
+class TSCaseCorrectingInsertHandler : InsertHandler<LookupElement> {
+    override fun handleInsert(context: InsertionContext, item: LookupElement) {
+        val elementBeingEdited = context.file.findElementAt(context.startOffset)
+        if (elementBeingEdited != null) {
+            // Remove the current element
+            context.editor.document.deleteString(elementBeingEdited.textOffset, context.editor.caretModel.currentCaret.offset)
+            // Insert the new element with proper casing
+            EditorModificationUtil.insertStringAtCaret(context.editor, item.lookupString, true)
+        }
+    }
+
+    companion object {
+        val INSTANCE = TSCaseCorrectingInsertHandler()
+    }
+}