+++ src/org/lukasj/idea/torquescript/completion/TSGlobalNSCallCompletionContributor.kt
@@ -1,0 +1,39 @@
+package org.lukasj.idea.torquescript.completion
+
+import com.intellij.codeInsight.completion.CompletionParameters
+import com.intellij.codeInsight.completion.CompletionProvider
+import com.intellij.codeInsight.completion.CompletionResultSet
+import com.intellij.codeInsight.lookup.LookupElementBuilder
+import com.intellij.util.ProcessingContext
+import org.lukasj.idea.torquescript.psi.impl.TSFunctionType
+import org.lukasj.idea.torquescript.reference.ReferenceUtil
+
+class TSGlobalNSCallCompletionContributor : CompletionProvider<CompletionParameters>() {
+    override fun addCompletions(
+        parameters: CompletionParameters,
+        context: ProcessingContext,
+        result: CompletionResultSet
+    ) {
+        val namespace = parameters
+            .position
+            .prevSibling
+            .prevSibling
+            .text
+
+        val project = parameters.originalFile.project
+
+        ReferenceUtil.getFunctions(parameters.position, project)
+            .filter { it.getFunctionType() != TSFunctionType.GLOBAL }
+            .filter { namespace != null && !it.getNamespace().equals(namespace, true) }
+            .forEach { function ->
+                result.addElement(
+                    LookupElementBuilder.create(function.name!!)
+                        .withCaseSensitivity(false)
+                        .withPresentableText("${function.getNamespace()}::${function.name}")
+                        .withTailText(function.getParameters().joinToString (prefix = "(", postfix = ")") { it.text })
+                        .withInsertHandler(TSCaseCorrectingInsertHandler.INSTANCE)
+                )
+            }
+    }
+
+}